+++ src/main/ts/Utils.ts
@@ -9,79 +9,10 @@
 import { Ref, watch, SetupContext } from 'vue';
 import { IPropTypes } from './components/EditorPropTypes';
 import type { Editor as HugeRTEEditor, EditorEvent } from 'hugerte';
-
-const validEvents = [
-  'onActivate',
-  'onAddUndo',
-  'onBeforeAddUndo',
-  'onBeforeExecCommand',
-  'onBeforeGetContent',
-  'onBeforeRenderUI',
-  'onBeforeSetContent',
-  'onBeforePaste',
-  'onBlur',
-  'onChange',
-  'onClearUndos',
-  'onClick',
-  'onContextMenu',
-  'onCommentChange',
-  'onCompositionEnd',
-  'onCompositionStart',
-  'onCompositionUpdate',
-  'onCopy',
-  'onCut',
-  'onDblclick',
-  'onDeactivate',
-  'onDirty',
-  'onDrag',
-  'onDragDrop',
-  'onDragEnd',
-  'onDragGesture',
-  'onDragOver',
-  'onDrop',
-  'onExecCommand',
-  'onFocus',
-  'onFocusIn',
-  'onFocusOut',
-  'onGetContent',
-  'onHide',
-  'onInit',
-  'onInput',
-  'onKeyDown',
-  'onKeyPress',
-  'onKeyUp',
-  'onLoadContent',
-  'onMouseDown',
-  'onMouseEnter',
-  'onMouseLeave',
-  'onMouseMove',
-  'onMouseOut',
-  'onMouseOver',
-  'onMouseUp',
-  'onNodeChange',
-  'onObjectResizeStart',
-  'onObjectResized',
-  'onObjectSelected',
-  'onPaste',
-  'onPostProcess',
-  'onPostRender',
-  'onPreProcess',
-  'onProgressState',
-  'onRedo',
-  'onRemove',
-  'onReset',
-  'onSaveContent',
-  'onSelectionChange',
-  'onSetAttrib',
-  'onSetContent',
-  'onShow',
-  'onSubmit',
-  'onUndo',
-  'onVisualAid'
-];
+import { validEvents } from '@hugerte/framework-integration-shared';
 
 const isValidKey = (key: string) =>
-  validEvents.map((event) => event.toLowerCase()).indexOf(key.toLowerCase()) !== -1;
+  validEvents.map((event) => 'on' + event.toLowerCase()).indexOf(key.toLowerCase()) !== -1;
 
 const bindHandlers = (initEvent: EditorEvent<any>, listeners: Record<string, any>, editor: HugeRTEEditor): void => {
   Object.keys(listeners)
@@ -89,10 +20,10 @@ const bindHandlers = (initEvent: EditorEvent<any>, listeners: Record<string, any
     .forEach((key: string) => {
       const handler = listeners[key];
       if (typeof handler === 'function') {
-        if (key === 'onInit') {
+        if (key === 'Init') {
           handler(initEvent, editor);
         } else {
-          editor.on(key.substring(2), (e: EditorEvent<any>) => handler(e, editor));
+          editor.on(key, (e: EditorEvent<any>) => handler(e, editor));
         }
       }
     });
@@ -127,41 +58,8 @@ const initEditor = (
   bindHandlers(initEvent, ctx.attrs, editor);
 };
 
-let unique = 0;
-
-const uuid = (prefix: string): string => {
-  const time = Date.now();
-  const random = Math.floor(Math.random() * 1000000000);
-
-  unique++;
-
-  return prefix + '_' + random + unique + String(time);
-};
-
-const isTextarea = (element: Element | null): element is HTMLTextAreaElement =>
-  element !== null && element.tagName.toLowerCase() === 'textarea';
-
-const normalizePluginArray = (plugins?: string | string[]): string[] => {
-  if (typeof plugins === 'undefined' || plugins === '') {
-    return [];
-  }
-
-  return Array.isArray(plugins) ? plugins : plugins.split(' ');
-};
-
-const mergePlugins = (initPlugins: string | string[] | undefined, inputPlugins?: string | string[]) =>
-  normalizePluginArray(initPlugins).concat(normalizePluginArray(inputPlugins));
-
-const isNullOrUndefined = (value: any): value is null | undefined =>
-  value === null || value === undefined;
-
 export {
   bindHandlers,
-  bindModelHandlers,
   initEditor,
-  isValidKey,
-  uuid,
-  isTextarea,
-  mergePlugins,
-  isNullOrUndefined
+  isValidKey, // TODO: Seems to only be used in a test, consider removing
 };